requirements.txt
Django==1.7.5
django-cors-headers==1.0.0
django-filter==0.9.2
django-geojson==2.7.0
django-leaflet==0.16.0
django-mssql==1.6.2
djangorestframework==3.1.1
gunicorn==19.3.0
jsonfield==1.0.3
mysql-connector-python==2.0.3
PyMySQL==0.6.6
pytz==2015.2
requests==2.5.3
six==1.9.0
